requirements.txt
from typing import runtime_checkable


torch
numpy
opencv-python
pillow
face_recognition
torchvision
runtime_checkable
